Prevent fruitless squeaks from Core X
This script appears to have been copied from `spawn_rnd_pickup`, as there is a remnant of the powerbomb section. As power bombs are not desired during situations when this is called, the decision has been made to exit from it early, so as to prevent Core X from emitting extra morph sounds when no pickup is spawned.